Fix peya

+++ dags/integrations/proc_peya_stock_integration.py
@@ -123,7 +123,7 @@ def _join_stock_and_promo_prices_from_s3(ds, ti):
         print(f"Number of records found on stock: {len(df.index)}")
 
         df.columns = map(str.upper, df.columns)
-        df["SKU"] = df["SKU"].astype("int64")
+        #df["SKU"] = df["SKU"].astype("int64")
         
         prev_exec_date = macros.ds_add(ds, -1).replace("-","/")
         prev_join_file_name = f"integraciones/last_millers/stock/out/peya/{prev_exec_date}/{peya_store_ids[store_id]}.csv"
@@ -230,7 +230,7 @@ def _join_stock_and_promo_prices_from_s3(ds, ti):
         print(f"Number of records found on stock: {len(df.index)}")
 
         df.columns = map(str.upper, df.columns)
-        df["SKU"] = df["SKU"].astype("int64")
+        #df["SKU"] = df["SKU"].astype("int64")
         
         prev_exec_date = macros.ds_add(ds, -1).replace("-","/")
         prev_join_file_name = f"integraciones/last_millers/promotions/out/peya/{prev_exec_date}/{peya_store_ids[store_id]}.csv"
